apps/api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
dspy-ai==2.6.2
python-dotenv==1.0.1 
sqlalchemy==2.0.23
psycopg[binary]==3.2.3 